Fix: fixed colors created middleware for verifying register and login created session guard hook
