+++ app/src/pages/Login.jsx
@@ -1,6 +1,6 @@
 import { Header } from "../components/layout/Header";
 import { Navbar } from "../components/layout/Navbar";
-import { Form } from "../components/shared/Form";
+import { LoginForm } from "../components/auth/LoginForm";
 import { PageLocation } from "../components/layout/PageLocation";
 import { PageHeading } from "../components/layout/PageHeading";
 import { Footer } from "../components/layout/Footer";
@@ -11,9 +11,9 @@ export const Login = () => {
       <Header />
       <Navbar />
       <div className="flex flex-col justify-center items-center min-h-[calc(100svh-172px)] bg-white dark:bg-black py-20 px-6">
-        <PageHeading page="Login"/>
+        <PageHeading page="Login" />
         <PageLocation pages={["Login"]} />
-        <Form isLogin={true} />
+        <LoginForm />
       </div>
       <Footer />
     </>
