+++ app/src/pages/Register.jsx
@@ -3,7 +3,7 @@ import { Navbar } from "../components/layout/Navbar";
 import { Footer } from "../components/layout/Footer";
 import { PageHeading } from "../components/layout/PageHeading";
 import { PageLocation } from "../components/layout/PageLocation";
-import { Form } from "../components/shared/Form";
+import { RegisterForm } from "../components/auth/RegisterForm";
 
 export const Register = () => {
   return (
@@ -11,9 +11,9 @@ export const Register = () => {
       <Header />
       <Navbar />
       <div className="flex flex-col justify-center items-center min-h-fullWithoutBars bg-white dark:bg-black py-20 px-6">
-        <PageHeading page="Registrati"/>
+        <PageHeading page="Registrati" />
         <PageLocation pages={["Registrati"]} />
-        <Form isLogin={false} />
+        <RegisterForm />
       </div>
       <Footer />
     </>
