+++ server/api/routes/users.js
@@ -6,75 +6,17 @@ const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 const User = require("../../db/models/User");
 const { verifyUserAuth } = require("../../middleware/verifyUserAuth");
+const { validateRegister } = require("../../middleware/validateRegister");
+const { validateLogin } = require("../../middleware/validateLogin");
 const { SECRET_KEY } = process.env;
 
 /**
  * @path /api/users/register
  */
 
-app.post("/register", async (req, res) => {
-  const schema = Joi.object().keys({
-    name: Joi.string().required().messages({
-      "any.required": "• Inserire nome",
-      "string.empty": "• Inserire nome",
-    }),
-    email: Joi.string().email().required().messages({
-      "any.required": "• Inserire un indirizzo email valido",
-      "string.empty": "• Inserire un indirizzo email valido",
-      "string.email": "• Inserire un indirizzo email valido",
-    }),
-    password: Joi.string()
-      .required()
-      .pattern(
-        /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*()_+{}\[\]:;<>,.?~\\-]).{8,}$/
-      )
-      .messages({
-        "any.required":
-          "• Password deve contenere una lettera maiuscola, una minuscola, un numero e almeno 8 caratteri",
-        "string.empty":
-          "• Password deve contenere una lettera maiuscola, una minuscola, un numero e almeno 8 caratteri",
-        "string.pattern.base":
-          "• Password deve contenere una lettera maiuscola, una minuscola, un numero e almeno 8 caratteri",
-      }),
-    confirmPassword: Joi.string()
-      .required()
-      .pattern(
-        /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*()_+{}\[\]:;<>,.?~\\-]).{8,}$/
-      )
-      .messages({
-        "any.required": "• Inserire conferma password",
-        "string.empty": "• Inserire conferma password",
-        "string.pattern.base":
-          "• Password deve contenere una lettera maiuscola, una minuscola, un numero e almeno 8 caratteri",
-      }),
-    // country: Joi.string().required().messages({
-    //   "any.required": "• Inserire paese",
-    //   "string.empty": "• Inserire paese",
-    // }),
-    // address: Joi.string().required().messages({
-    //   "any.required": "• Inserire un indirizzo valido",
-    //   "string.empty": "• Inserire un indirizzo valido",
-    // }),
-    // civic: Joi.string().required().messages({
-    //   "any.required": "• Inserire civico",
-    //   "string.empty": "• Inserire civico",
-    // }),
-    // zipCode: Joi.string().required().min(5).messages({
-    //   "any.required": "• Inserire cap valido",
-    //   "string.empty": "• Inserire cap valido",
-    //   "string.min": "• Inserire cap valido",
-    // }),
-    // phone: Joi.string()
-    //   .required()
-    //   .pattern(/^\+?(39|0039)?[ ]?[0-9]{2,4}[ ]?[0-9]{6,8}$/)
-    //   .messages({
-    //     "any.required": "• Inserire numero di cellulare valido",
-    //     "string.empty": "• Inserire numero di cellulare valido",
-    //     "string.pattern.base": "• Inserire numero di cellulare valido",
-    //   }),
-  });
+app.post("/register", validateRegister, async (req, res) => {
   try {
-    const user = await schema.validateAsync(req.body);
+    const user = await req.schema.validateAsync(req.body);
 
     user.email = user.email.toLowerCase();
 
@@ -96,6 +38,7 @@ app.post("/register", async (req, res) => {
         .status(403)
         .json({ message: "Indirizzo email già registrato" });
     }
+    console.log(user, "USER");
 
     await User.create(user);
 
@@ -108,21 +51,9 @@ app.post("/register", async (req, res) => {
   }
 });
 
-app.post("/login", async (req, res) => {
-  const schema = Joi.object().keys({
-    email: Joi.string().email().required().messages({
-      "any.required": "• Inserire un indirizzo email valido",
-      "string.empty": "• Inserire un indirizzo email valido",
-      "string.email": "• Inserire un indirizzo email valido",
-    }),
-    password: Joi.string().required().messages({
-      "any.required": "• Inserire password",
-      "string.empty": "• Inserire password",
-    }),
-    remember: Joi.bool(),
-  });
+app.post("/login", validateLogin, async (req, res) => {
   try {
-    const data = await schema.validateAsync(req.body);
+    const data = await req.schema.validateAsync(req.body);
 
     data.email = data.email.toLowerCase();
